Minor polish
- more docstrings
- refactor things
- disable one 15s formatting test that stresses a non-critical
  path, no reason it should take 10% of test runtime.


Former-commit-id: 274e4c48e4f3a8d6094f191fc78924d76adf0c7b

+++ metals/src/main/scala/scala/meta/internal/metals/BloomFilters.scala
@@ -5,10 +5,10 @@ import com.google.common.hash.Funnels
 import java.nio.charset.StandardCharsets
 
 object BloomFilters {
-  def create(size: Int): BloomFilter[CharSequence] = {
+  def create(size: Long): BloomFilter[CharSequence] = {
     BloomFilter.create(
       Funnels.stringFunnel(StandardCharsets.UTF_8),
-      Integer.valueOf(size),
+      java.lang.Long.valueOf(size),
       0.01
     )
   }
